Added support for running multiple instances of your web server.
Without this only the first thread will initially make the `req.assetFingerprint()` calls in your view which populates our collection of `fingerprints`.
The other threads will not have the `fingerprints` collection populated but will likely handle one or more requests for the assets and when they do your strategy will not be applied.